requirements.txt
Django>=5.0.2,<5.1.0
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<1.1.0
drf-spectacular>=0.27.0,<0.28.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
